--- src/renderer/const.test.ts
@@ -1,18 +1,0 @@
-import { AssetBNB } from '@thorchain/asgardex-util'
-
-import { AssetBUSDBAF, AssetBUSDBD1, isPricePoolAsset } from './const'
-
-describe('view/pools/types', () => {
-  describe('isPricePoolAsset', () => {
-    it('returns true for BUSDB', () => {
-      expect(isPricePoolAsset(AssetBUSDBAF)).toBeTruthy()
-      expect(isPricePoolAsset(AssetBUSDBD1)).toBeTruthy()
-    })
-    it('returns false for BNB', () => {
-      expect(isPricePoolAsset(AssetBNB)).toBeFalsy()
-    })
-    it('returns false for deprecated asset ', () => {
-      expect(isPricePoolAsset({ chain: 'BNB', symbol: 'RUNE-1AF', ticker: 'RUNE' })).toBeFalsy()
-    })
-  })
-})
