+++ src/renderer/services/midgard/pools.ts
@@ -9,8 +9,8 @@ import * as Rx from 'rxjs'
 import { combineLatest } from 'rxjs'
 import { catchError, map, retry, shareReplay, startWith, switchMap, filter } from 'rxjs/operators'
 
-import { isPricePoolAsset, ONE_BN, PRICE_POOLS_WHITELIST } from '../../const'
-import { getRuneAsset } from '../../helpers/assetHelper'
+import { ONE_BN, PRICE_POOLS_WHITELIST } from '../../const'
+import { getRuneAsset, isPricePoolAsset } from '../../helpers/assetHelper'
 import { sequenceTOption } from '../../helpers/fpHelpers'
 import { LiveData, liveData } from '../../helpers/rx/liveData'
 import { observableState, triggerStream } from '../../helpers/stateHelper'
